+++ packages/component-better-link/src/betterLinkDocumentMod.ariaLabel.spec.ts
@@ -1,12 +1,10 @@
 /** @jest-environment @happy-dom/jest-environment */
 /// <reference types="jest" />
 
-import {
-  parseDocumentFragmentFromString,
-  serializeDocumentFragmentIntoString
-} from 'botframework-webchat-component/internal.js';
 import { micromark } from 'micromark';
 import betterLinkDocumentMod, { type BetterLinkDocumentModDecoration } from './betterLinkDocumentMod';
+import parseDocumentFragmentFromString from './parseDocumentFragmentFromString';
+import serializeDocumentFragmentIntoString from './serializeDocumentFragmentIntoString';
 
 const BASE_MARKDOWN = '[Example](https://example.com)';
 let baseHTML: string;
@@ -24,7 +22,7 @@ describe('When passing "ariaLabel" option with "Hello, World!"', () => {
   });
 
   test('should have "aria-label" attribute set to "Hello, World!"', () =>
-    expect(actual.querySelector('a').getAttribute('aria-label')).toBe('Hello, World!'));
+    expect(actual.querySelector('a')?.getAttribute('aria-label')).toBe('Hello, World!'));
 
   test('should match snapshot', () =>
     expect(serializeDocumentFragmentIntoString(actual)).toBe(
@@ -44,7 +42,7 @@ describe('When passing "ariaLabel" option with false', () => {
   });
 
   test('should have "aria-label" attribute removed', () =>
-    expect(actual.querySelector('a').hasAttribute('aria-label')).toBe(false));
+    expect(actual.querySelector('a')?.hasAttribute('aria-label')).toBe(false));
 
   test('should match snapshot', () =>
     expect(serializeDocumentFragmentIntoString(actual)).toBe(