+++ packages/component-better-link/src/betterLinkDocumentMod.iconClassName.spec.ts
@@ -1,15 +1,13 @@
 /** @jest-environment @happy-dom/jest-environment */
 /// <reference types="jest" />
 
-import {
-  parseDocumentFragmentFromString,
-  serializeDocumentFragmentIntoString
-} from 'botframework-webchat-component/internal.js';
 import { micromark } from 'micromark';
 import betterLinkDocumentMod, { type BetterLinkDocumentModDecoration } from './betterLinkDocumentMod';
+import parseDocumentFragmentFromString from './parseDocumentFragmentFromString';
+import serializeDocumentFragmentIntoString from './serializeDocumentFragmentIntoString';
 
 const BASE_MARKDOWN = '[Example](https://example.com)';
-let baseHTML;
+let baseHTML: string;
 
 beforeEach(() => {
   baseHTML = micromark(BASE_MARKDOWN, { allowDangerousHtml: true });
@@ -24,18 +22,18 @@ describe('When passing "iconAlt" option with "Hello, World!" and "iconClassName"
   });
 
   test('should have icon image with "alt" attribute set to empty string', () =>
-    expect(actual.querySelector('img').getAttribute('alt')).toBe(''));
+    expect(actual.querySelector('img')?.getAttribute('alt')).toBe(''));
 
   test('should have icon image with "class" attribute set to "my-icon"', () =>
-    expect(actual.querySelector('img').classList.contains('my-icon')).toBe(true));
+    expect(actual.querySelector('img')?.classList.contains('my-icon')).toBe(true));
 
   test('should have icon image with "src" attribute set to a transparent GIF', () =>
-    expect(actual.querySelector('img').getAttribute('src')).toBe(
+    expect(actual.querySelector('img')?.getAttribute('src')).toBe(
       'data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7'
     ));
 
   test('should have icon image with "title" attribute set to "Hello, World!"', () =>
-    expect(actual.querySelector('img').getAttribute('title')).toBe('Hello, World!'));
+    expect(actual.querySelector('img')?.getAttribute('title')).toBe('Hello, World!'));
 
   test('should match snapshot', () =>
     expect(serializeDocumentFragmentIntoString(actual)).toBe(