+++ packages/component-better-link/src/betterLinkDocumentMod.target.spec.ts
@@ -1,12 +1,10 @@
 /** @jest-environment @happy-dom/jest-environment */
 /// <reference types="jest" />
 
-import {
-  parseDocumentFragmentFromString,
-  serializeDocumentFragmentIntoString
-} from 'botframework-webchat-component/internal.js';
 import { micromark } from 'micromark';
 import betterLinkDocumentMod, { type BetterLinkDocumentModDecoration } from './betterLinkDocumentMod';
+import serializeDocumentFragmentIntoString from './serializeDocumentFragmentIntoString';
+import parseDocumentFragmentFromString from './parseDocumentFragmentFromString';
 
 const BASE_MARKDOWN = '[Example](https://example.com)';
 let baseHTML: string;
@@ -15,36 +13,36 @@ beforeEach(() => {
   baseHTML = micromark(BASE_MARKDOWN, { allowDangerousHtml: true });
 });
 
-describe('When passing "title" option with "Hello, World!"', () => {
+describe('When passing "target" option with "noopener noreferer"', () => {
   let actual: DocumentFragment;
-  const decoration: BetterLinkDocumentModDecoration = { title: 'Hello, World!' };
+  const decoration: BetterLinkDocumentModDecoration = { target: 'noopener noreferer' };
 
   beforeEach(() => {
     actual = betterLinkDocumentMod(parseDocumentFragmentFromString(baseHTML), () => decoration);
   });
 
-  test('should have "title" attribute set to "Hello, World!"', () =>
-    expect(actual.querySelector('a').getAttribute('title')).toBe('Hello, World!'));
+  test('should have "target" attribute set to "noopener noreferer"', () =>
+    expect(actual.querySelector('a')?.getAttribute('target')).toBe('noopener noreferer'));
 
   test('should match snapshot', () =>
     expect(serializeDocumentFragmentIntoString(actual)).toBe(
-      '<p xmlns="http://www.w3.org/1999/xhtml"><a href="https://example.com" title="Hello, World!">Example</a></p>'
+      '<p xmlns="http://www.w3.org/1999/xhtml"><a href="https://example.com" target="noopener noreferer">Example</a></p>'
     ));
 });
 
-describe('When passing "title" option with false', () => {
+describe('When passing "target" option with false', () => {
   let actual: DocumentFragment;
-  const decoration: BetterLinkDocumentModDecoration = { title: false };
+  const decoration: BetterLinkDocumentModDecoration = { target: false };
 
   beforeEach(() => {
     actual = betterLinkDocumentMod(
-      parseDocumentFragmentFromString('<a href="https://example.com" title="Hello, World!">Example</a>'),
+      parseDocumentFragmentFromString('<a href="https://example.com" target="noopener noreferer">Example</a>'),
       () => decoration
     );
   });
 
-  test('should have "title" attribute removed', () =>
-    expect(actual.querySelector('a').hasAttribute('title')).toBe(false));
+  test('should have "target" attribute removed', () =>
+    expect(actual.querySelector('a')?.hasAttribute('target')).toBe(false));
 
   test('should match snapshot', () =>
     expect(serializeDocumentFragmentIntoString(actual)).toBe(