+++ packages/component-better-link/src/index.ts
@@ -1,0 +1,8 @@
+export {
+  default as betterLinkDocumentMod,
+  type BetterLinkDocumentModAttributeSetter,
+  type BetterLinkDocumentModDecoration
+} from './betterLinkDocumentMod';
+export { default as parseDocumentFragmentFromString } from './parseDocumentFragmentFromString';
+export { default as serializeDocumentFragmentIntoString } from './serializeDocumentFragmentIntoString';
+export { default as stripParagraphContainer } from './stripParagraphContainer';