+++ packages/component-better-link/src/stripParagraphContainer.ts
@@ -1,0 +1,6 @@
+export default function stripParagraphContainer(html: string): string {
+  return html
+    .trim()
+    .replace(/^<p\sxmlns="http:\/\/www.w3.org\/1999\/xhtml">/u, '')
+    .replace(/<\/p>$/u, '');
+}