+++ packages/component/src/hooks/internal/useInternalMarkdownIt.js
@@ -1,7 +1,0 @@
-import useWebChatUIContext from './useWebChatUIContext';
-
-function useInternalMarkdownIt() {
-  return useWebChatUIContext().internalMarkdownItState;
-}
-
-export default useInternalMarkdownIt;